upd - Removed Extensification dependency
We've removed a dependency on Extensification

---

This project no longer depends on Extensification because it's deprecated. This is a commit to increment the version of the project and to remove this library as a dependency. More information about this deprecation can be found on our blog.

---

Type: upd
Breaking: False
Doc Required: False
Part: 1/1